+++ src/gateway/chat-auth-readiness.integration.test.ts
@@ -1,18 +1,23 @@
+/* @vitest-environment jsdom */
+
 import { expect, it } from "vitest";
-import { getRuntimeConfig } from "../../../../src/config/io.js";
-import { resolveSessionStorePathCore } from "../../../../src/config/sessions.js";
-import { upsertSessionEntryCore } from "../../../../src/config/sessions/session-accessor.js";
-import { createLifecycleEventBroadcastHandler } from "../../../../src/gateway/server-session-events.js";
-import { loadGatewaySessionEntryReadOnly } from "../../../../src/gateway/session-utils.js";
-import { applySessionModelSelection } from "../../../../src/model-picker/apply-session-model-selection.js";
-import { onSessionLifecycleEvent } from "../../../../src/sessions/session-lifecycle-events.js";
-import { withOpenClawTestState } from "../../../../src/test-utils/openclaw-test-state.js";
-import { createTestGatewayClient } from "../../test-helpers/gateway-client.ts";
-import { waitForFast } from "../../test-helpers/wait-for.ts";
-import { makeChatHost, makeRequestMock } from "./chat-host.test-support.ts";
-import { handlePageGatewayEvent } from "./chat-state-events.ts";
-import type { ChatPageHost } from "./chat-state-host.ts";
-import { refreshChatMetadata, retireChatMetadataRequests } from "./chat-state-refresh.ts";
+import { makeChatHost, makeRequestMock } from "../../ui/src/pages/chat/chat-host.test-support.ts";
+import { handlePageGatewayEvent } from "../../ui/src/pages/chat/chat-state-events.ts";
+import type { ChatPageHost } from "../../ui/src/pages/chat/chat-state-host.ts";
+import {
+  refreshChatMetadata,
+  retireChatMetadataRequests,
+} from "../../ui/src/pages/chat/chat-state-refresh.ts";
+import { createTestGatewayClient } from "../../ui/src/test-helpers/gateway-client.ts";
+import { waitForFast } from "../../ui/src/test-helpers/wait-for.ts";
+import { getRuntimeConfig } from "../config/io.js";
+import { resolveSessionStorePathCore } from "../config/sessions.js";
+import { upsertSessionEntryCore } from "../config/sessions/session-accessor.js";
+import { applySessionModelSelection } from "../model-picker/apply-session-model-selection.js";
+import { onSessionLifecycleEvent } from "../sessions/session-lifecycle-events.js";
+import { withOpenClawTestState } from "../test-utils/openclaw-test-state.js";
+import { createLifecycleEventBroadcastHandler } from "./server-session-events.js";
+import { loadGatewaySessionEntryReadOnly } from "./session-utils.js";
 
 it("refreshes a retained pane from a persisted profile-only selection through the Gateway lifecycle broadcaster", async () => {
   await withOpenClawTestState({ scenario: "minimal" }, async () => {
