+++ src/gateway/server-methods/chat-metadata-runtime.test.ts
@@ -1,187 +1,20 @@
 import { describe, expect, test, vi } from "vitest";
 import { createDeferred } from "../../../test/helpers/promise.js";
-import {
-  resolveUsableAgentCredentialModes,
-  type AgentCredentialMap,
-} from "../../agents/agent-auth-credentials.js";
+import type { AgentCredentialMap } from "../../agents/agent-auth-credentials.js";
 import type { AuthProfileStore } from "../../agents/auth-profiles.js";
-import type { ModelCatalogEntry, ModelCatalogSnapshot } from "../../agents/model-catalog.types.js";
-import {
-  setPreparedModelFullCatalogAuth,
-  setPreparedModelRuntimeAuthStore,
-} from "../../agents/prepared-model-runtime-auth.js";
+import type { ModelCatalogSnapshot } from "../../agents/model-catalog.types.js";
+import { setPreparedModelFullCatalogAuth } from "../../agents/prepared-model-runtime-auth.js";
 import { markPreparedModelCatalogFull } from "../../agents/prepared-model-runtime.full-catalog.js";
-import type { PreparedModelRuntimeSnapshot } from "../../agents/prepared-model-runtime.js";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
-import { createGatewayChatMetadataRuntime } from "./chat-metadata-runtime.js";
-import type { GatewayRequestContext } from "./types.js";
-
-function createOwner(
-  config: OpenClawConfig,
-  id: string,
-  credentials: AgentCredentialMap = {},
-  provider = "test",
-  api?: ModelCatalogEntry["api"],
-): PreparedModelRuntimeSnapshot {
-  const model = { id, name: id, provider, ...(api ? { api } : {}) };
-  const authStore: AuthProfileStore = {
-    version: 1,
-    profiles: Object.fromEntries(
-      Object.entries(credentials).map(([credentialProvider, credential]) => [
-        `${credentialProvider}:prepared`,
-        { ...credential, provider: credentialProvider },
-      ]),
-    ),
-  };
-  const owner: PreparedModelRuntimeSnapshot = {
-    catalogOwner: { agentId: "main", workspaceDir: `/tmp/${id}/workspace` },
-    agentId: "main",
-    agentDir: `/tmp/${id}/agent`,
-    workspaceDir: `/tmp/${id}/workspace`,
-    activeProjectKeys: [],
-    config,
-    authModes: resolveUsableAgentCredentialModes(credentials),
-    metadataSnapshot: { index: { plugins: [] }, plugins: [] } as never,
-    allowGatewaySubagentBinding: false,
-    modelCatalog: {
-      entries: [model],
-      routeVariants: api ? [model] : [],
-    },
-    configuredRuntimeModels: [],
-    inlineProviderModels: [],
-    createStores: () => ({
-      authStorage: { getAll: () => credentials } as never,
-      modelRegistry: {} as never,
-    }),
-  };
-  setPreparedModelRuntimeAuthStore(owner, authStore);
-  return owner;
-}
-
-function createHarness(
-  initialConfig: OpenClawConfig = { agents: { list: [{ id: "main", default: true }] } },
-  runtimeOptions: {
-    beforeRefresh?: () => Promise<void>;
-    refreshOnRead?: boolean;
-    useDefaultProjection?: boolean;
-    onChanged?: () => void;
-  } = {},
-) {
-  const { useDefaultProjection = false, ...gatewayRuntimeOptions } = runtimeOptions;
-  let config = initialConfig;
-  let owner = createOwner(config, "first");
-  let skillsVersion = 1;
-  let pluginRegistryVersion = 1;
-  let authStore: AuthProfileStore | undefined = { version: 1, profiles: {} };
-  let authStoreRevision = 1;
-  const invalidProjections = new WeakSet<object>();
-  const getPreparedOwner = vi.fn((): PreparedModelRuntimeSnapshot | undefined => owner);
-  const getPreparedAuthStore = vi.fn(() => authStore);
-  const getAuthStoreRevision = vi.fn(() => authStoreRevision);
-  const getSkillsVersion = vi.fn(() => skillsVersion);
-  const getPluginRegistryVersion = vi.fn(() => pluginRegistryVersion);
-  const buildCommands = vi.fn(async () => ({
-    commands: [{ name: `command-${skillsVersion}-${pluginRegistryVersion}` }],
-  }));
-  const buildProjection = vi.fn(
-    async ({
-      facts,
-    }: {
-      facts: {
-        authStore: AuthProfileStore;
-        modelCatalog: ModelCatalogSnapshot;
-        owner: PreparedModelRuntimeSnapshot;
-      };
-    }) => {
-      const modelCatalog = facts.modelCatalog;
-      return {
-        modelCatalog: modelCatalog.entries,
-        models: modelCatalog.entries,
-      };
-    },
-  );
-  const context = {
-    getRuntimeConfig: () => config,
-    loadGatewayModelCatalogSnapshot: async (params?: { readOnly?: boolean }) => {
-      const modelCatalog =
-        params?.readOnly === false && owner.loadFullModelCatalog
-          ? await owner.loadFullModelCatalog()
-          : owner.modelCatalog;
-      return {
-        ...modelCatalog,
-        agentId: owner.agentId,
-        agentDir: owner.agentDir,
-        workspaceDir: owner.workspaceDir,
-        config: owner.config,
-      };
-    },
-    logGateway: {
-      debug: vi.fn(),
-      info: vi.fn(),
-      warn: vi.fn(),
-      error: vi.fn(),
-    },
-  } as unknown as GatewayRequestContext;
-  const runtime = createGatewayChatMetadataRuntime({
-    getConfig: () => config,
-    getContext: () => context,
-    log: context.logGateway,
-    ...gatewayRuntimeOptions,
-    deps: {
-      getPreparedOwner,
-      getPreparedAuthStore,
-      getAuthStoreRevision,
-      getSkillsVersion,
-      getPluginRegistryVersion,
-      buildCommands,
-      ...(useDefaultProjection
-        ? {}
-        : {
-            buildProjection: async (params) => {
-              const projection = await buildProjection(params);
-              return {
-                read: () => projection,
-                isCurrent: () => !invalidProjections.has(projection),
-              };
-            },
-          }),
-    },
-  });
-  return {
-    buildCommands,
-    buildProjection,
-    getPluginRegistryVersion,
-    getAuthStoreRevision,
-    getPreparedAuthStore,
-    getPreparedOwner,
-    getSkillsVersion,
-    invalidProjections,
-    runtime,
-    setConfig(next: OpenClawConfig) {
-      config = next;
-    },
-    setAuthStore(next: AuthProfileStore | undefined) {
-      authStore = next;
-    },
-    setAuthStoreRevision(next: number) {
-      authStoreRevision = next;
-    },
-    setOwner(next: PreparedModelRuntimeSnapshot) {
-      owner = next;
-    },
-    setPluginRegistryVersion(next: number) {
-      pluginRegistryVersion = next;
-    },
-    setSkillsVersion(next: number) {
-      skillsVersion = next;
-    },
-  };
-}
+import {
+  createChatMetadataHarness,
+  createChatMetadataOwner,
+} from "./chat-metadata-runtime.test-support.js";
 
 describe("gateway chat metadata runtime", () => {
   test("notifies once per settlement, including same-epoch recovery, without an unavailable-read loop", async () => {
     const onChanged = vi.fn();
-    const harness = createHarness(undefined, { onChanged });
+    const harness = createChatMetadataHarness(undefined, { onChanged });
     const owner = harness.getPreparedOwner();
     harness.getPreparedOwner.mockReturnValue(undefined);
     await expect(harness.runtime.refresh()).rejects.toThrow("owner is unavailable");
@@ -205,7 +38,7 @@ describe("gateway chat metadata runtime", () => {
     "never announces an obsolete build's %s after terminal failure",
     async (settlement) => {
       const onChanged = vi.fn();
-      const harness = createHarness(undefined, { onChanged });
+      const harness = createChatMetadataHarness(undefined, { onChanged });
       const gate = createDeferred();
       harness.buildProjection.mockImplementationOnce(async ({ facts }) => {
         await gate.promise;
@@ -231,7 +64,7 @@ describe("gateway chat metadata runtime", () => {
 
   test("refreshes lazily on the first read when configured", async () => {
     const beforeRefresh = vi.fn(async () => {});
-    const harness = createHarness(undefined, { beforeRefresh, refreshOnRead: true });
+    const harness = createChatMetadataHarness(undefined, { beforeRefresh, refreshOnRead: true });
 
     expect(harness.buildProjection).not.toHaveBeenCalled();
     await expect(harness.runtime.read({ agentId: "main" })).resolves.toMatchObject({
@@ -243,7 +76,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("single-flights equivalent refreshes and reads", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     const releaseModels = createDeferred();
     harness.buildProjection.mockImplementationOnce(async ({ facts }) => {
       await releaseModels.promise;
@@ -271,7 +104,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("serves published metadata without request-time generation reads", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
     harness.getPreparedOwner.mockClear();
     harness.getPreparedAuthStore.mockClear();
@@ -291,7 +124,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("serves startup projections from the published generation", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
     harness.getPreparedOwner.mockClear();
     harness.getPreparedAuthStore.mockClear();
@@ -319,7 +152,7 @@ describe("gateway chat metadata runtime", () => {
   test("keeps large-roster neutral projections prepared outside the session cache", async () => {
     const defaultAgentId = "agent-0";
     const agentIds = Array.from({ length: 65 }, (_, index) => `agent-${index}`);
-    const harness = createHarness({
+    const harness = createChatMetadataHarness({
       agents: {
         list: agentIds.map((id) => ({
           id,
@@ -352,7 +185,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("caches a session auth projection separately from the neutral projection", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
 
     const sessionEntry = {
@@ -379,7 +212,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("ready reads never prepare or await a cold or pending exact profile", async () => {
-    const harness = createHarness(undefined, { refreshOnRead: true });
+    const harness = createChatMetadataHarness(undefined, { refreshOnRead: true });
     const sessionEntry = {
       authProfileOverride: "test:session",
       authProfileOverrideSource: "user" as const,
@@ -428,7 +261,7 @@ describe("gateway chat metadata runtime", () => {
   test.each(["neutral", "profile"] as const)(
     "ready reads omit an invalid %s wrapper until a canonical read refreshes it",
     async (invalid) => {
-      const harness = createHarness();
+      const harness = createChatMetadataHarness();
       const sessionEntry = {
         authProfileOverride: "test:session",
         authProfileOverrideSource: "user" as const,
@@ -468,7 +301,7 @@ describe("gateway chat metadata runtime", () => {
   test.each(["invalidate", "pending refresh", "failed", "stale facts"] as const)(
     "ready reads omit projections after %s without starting replacement work",
     async (state) => {
-      const harness = createHarness(undefined, { refreshOnRead: true });
+      const harness = createChatMetadataHarness(undefined, { refreshOnRead: true });
       const sessionEntry = { authProfileOverride: "test:session" };
       await harness.runtime.refresh();
       await harness.runtime.readStartup({ agentId: "main", sessionEntry });
@@ -510,7 +343,7 @@ describe("gateway chat metadata runtime", () => {
   test.each(["resolve", "reject"] as const)(
     "an evicted profile's late %s cannot replace or delete its newer ready entry",
     async (settlement) => {
-      const harness = createHarness();
+      const harness = createChatMetadataHarness();
       await harness.runtime.refresh();
       const sessionEntry = { authProfileOverride: "test:evicted" };
       const release = createDeferred();
@@ -569,7 +402,7 @@ describe("gateway chat metadata runtime", () => {
       locked: false,
     },
   ])("projects $name provenance", async ({ sessionEntry, locked }) => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
 
     await harness.runtime.readStartup({
@@ -595,7 +428,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("reuses the prepared generation for an equivalent config replacement", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
     const first = await harness.runtime.read({ agentId: "main" });
 
@@ -611,7 +444,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("rebuilds after an auth store publishes a newer revision", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     harness.setAuthStore(undefined);
     harness.buildProjection.mockImplementation(async ({ facts }) => ({
       modelCatalog: facts.owner.modelCatalog.entries,
@@ -642,9 +475,9 @@ describe("gateway chat metadata runtime", () => {
   test.each(["before", "after"] as const)(
     "converges to models.list availability when owner auth publishes %s attachment",
     async (publicationOrder) => {
-      const harness = createHarness(undefined, { useDefaultProjection: true });
+      const harness = createChatMetadataHarness(undefined, { useDefaultProjection: true });
       harness.setAuthStore({ version: 1, profiles: {} });
-      const preparedOwner = createOwner(
+      const preparedOwner = createChatMetadataOwner(
         harness.getPreparedOwner()!.config,
         "gpt-5.4",
         {
@@ -686,7 +519,7 @@ describe("gateway chat metadata runtime", () => {
         list: [{ id: "main", default: true }],
       },
     } as OpenClawConfig;
-    const harness = createHarness(config, { useDefaultProjection: true });
+    const harness = createChatMetadataHarness(config, { useDefaultProjection: true });
     const credentials: AgentCredentialMap = {
       openai: {
         type: "oauth",
@@ -695,7 +528,7 @@ describe("gateway chat metadata runtime", () => {
         expires: Date.now() + 30 * 60_000,
       },
     };
-    const owner = createOwner(
+    const owner = createChatMetadataOwner(
       config,
       "gpt-5.6-sol",
       credentials,
@@ -733,7 +566,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("keeps a locked session unavailable while the neutral prepared route is usable", async () => {
-    const harness = createHarness(
+    const harness = createChatMetadataHarness(
       {
         agents: {
           defaults: {
@@ -746,7 +579,7 @@ describe("gateway chat metadata runtime", () => {
       { useDefaultProjection: true },
     );
     harness.setOwner(
-      createOwner(
+      createChatMetadataOwner(
         harness.getPreparedOwner()!.config,
         "gpt-5.6-sol",
         {
@@ -796,8 +629,8 @@ describe("gateway chat metadata runtime", () => {
         expires: Date.now() + 30 * 60_000,
       },
     };
-    const harness = createHarness(config, { useDefaultProjection: true });
-    const owner = createOwner(
+    const harness = createChatMetadataHarness(config, { useDefaultProjection: true });
+    const owner = createChatMetadataOwner(
       config,
       "gpt-5.6-sol",
       credentials,
@@ -853,7 +686,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("rejects a completed catalog without its matching prepared auth generation", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     const owner = harness.getPreparedOwner()!;
     const unpairedCatalog = markPreparedModelCatalogFull({ ...owner.modelCatalog });
     harness.setOwner({ ...owner, readFullModelCatalog: () => unpairedCatalog });
@@ -864,7 +697,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("retains a generation while auth store revisions are unchanged", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     harness.getPreparedAuthStore.mockImplementation(() => ({ version: 1, profiles: {} }));
     await harness.runtime.refresh();
     const first = await harness.runtime.read({ agentId: "main" });
@@ -879,7 +712,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("refreshes config, catalog-auth, skills, and plugin generations", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
     const first = await harness.runtime.read({ agentId: "main" });
 
@@ -898,7 +731,7 @@ describe("gateway chat metadata runtime", () => {
       tools: { swarm: { enabled: true } },
     };
     harness.setConfig(nextConfig);
-    harness.setOwner(createOwner(nextConfig, "second"));
+    harness.setOwner(createChatMetadataOwner(nextConfig, "second"));
     harness.runtime.invalidate();
     await harness.runtime.refresh();
     const configAndOwnerChanged = await harness.runtime.read({ agentId: "main" });
@@ -914,7 +747,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("waits for replacement only for canonical metadata and session auth projections", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
 
     harness.runtime.invalidate();
@@ -954,7 +787,7 @@ describe("gateway chat metadata runtime", () => {
       tools: { swarm: { enabled: true } },
     };
     harness.setConfig(nextConfig);
-    harness.setOwner(createOwner(nextConfig, "replacement"));
+    harness.setOwner(createChatMetadataOwner(nextConfig, "replacement"));
     await harness.runtime.refresh();
 
     await expect(read).resolves.toMatchObject({
@@ -971,7 +804,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("retries a session projection invalidated while it is awaiting", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
     const releaseProjection = createDeferred();
     harness.buildProjection.mockImplementationOnce(async ({ facts }) => {
@@ -996,7 +829,7 @@ describe("gateway chat metadata runtime", () => {
       tools: { swarm: { enabled: true } },
     };
     harness.setConfig(nextConfig);
-    harness.setOwner(createOwner(nextConfig, "replacement"));
+    harness.setOwner(createChatMetadataOwner(nextConfig, "replacement"));
     harness.runtime.invalidate();
     await harness.runtime.refresh();
 
@@ -1008,7 +841,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("discards a projection failure from an invalidated generation", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
     const releaseProjection = createDeferred();
     harness.buildProjection.mockImplementationOnce(async () => {
@@ -1030,7 +863,7 @@ describe("gateway chat metadata runtime", () => {
       tools: { swarm: { enabled: true } },
     };
     harness.setConfig(nextConfig);
-    harness.setOwner(createOwner(nextConfig, "replacement"));
+    harness.setOwner(createChatMetadataOwner(nextConfig, "replacement"));
     harness.runtime.invalidate();
     await harness.runtime.refresh();
 
@@ -1042,7 +875,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("resolves the replacement gate after a coalesced second invalidation", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
     const releaseCommands = createDeferred();
     harness.buildCommands.mockImplementationOnce(async () => {
@@ -1070,7 +903,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("retries an unavailable owner on the next read once it is published again", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
 
     harness.getPreparedOwner.mockReturnValue(undefined);
@@ -1081,7 +914,7 @@ describe("gateway chat metadata runtime", () => {
       'prepared chat metadata owner is unavailable for agent "main"',
     );
 
-    const recovered = createOwner(
+    const recovered = createChatMetadataOwner(
       { agents: { list: [{ id: "main", default: true }] } },
       "recovered",
     );
@@ -1094,7 +927,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("rejects replacement waiters on failure and recovers on a later generation", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     await harness.runtime.refresh();
 
     harness.runtime.invalidate();
@@ -1105,7 +938,7 @@ describe("gateway chat metadata runtime", () => {
 
     const nextConfig = { agents: { list: [{ id: "main", default: true }] } };
     harness.setConfig(nextConfig);
-    harness.setOwner(createOwner(nextConfig, "recovered"));
+    harness.setOwner(createChatMetadataOwner(nextConfig, "recovered"));
     harness.runtime.invalidate();
     await harness.runtime.refresh();
 
@@ -1115,7 +948,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("omits failed command preparation without losing models", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     harness.buildCommands.mockRejectedValueOnce(new Error("skill scan failed"));
 
     await harness.runtime.refresh();
@@ -1126,7 +959,7 @@ describe("gateway chat metadata runtime", () => {
   });
 
   test("does not publish a generation whose neutral projection failed", async () => {
-    const harness = createHarness();
+    const harness = createChatMetadataHarness();
     harness.buildProjection.mockRejectedValueOnce(new Error("startup projection failed"));
 
     await expect(harness.runtime.refresh()).rejects.toThrow("startup projection failed");
